+++ src/methods/$split-with/string-split-with.mjs
@@ -1,6 +1,6 @@
 function findNextIndex(str, predicate, start = 0) {
   for (let i = start; i < str.length; i++) {
-    if (predicate(str[i])) {
+    if (predicate(str[i], i)) {
       return i;
     }
   }
@@ -8,6 +8,11 @@ function findNextIndex(str, predicate, start = 0) {
 }
 
 function* stringSplitWith(str, predicate) {
+  if (predicate instanceof RegExp) {
+    yield* str.split(predicate);
+    return;
+  }
+
   let matchIndex = -1;
   let nextMatchIndex;
   while ((nextMatchIndex = findNextIndex(str, predicate, matchIndex + 1)) >= 0) {